Auditdom: move code away from CLI module
